+++ package.json
@@ -23,10 +23,10 @@
     "eslint-plugin-node": "^11.1.0",
     "eslint-plugin-react": "^7.33.2",
     "js-cookie": "^3.0.5",
-    "next": "12.1.0",
+    "next": "^12.2.0",
     "react": "^17.0.2",
     "react-cookie": "^7.1.4",
-    "react-dom": "17.0.2",
+    "react-dom": "^17.0.2",
     "react-query": "^3.39.3",
     "react-router-dom": "^6.15.0",
     "recoil": "^0.7.7",
